enhanced-rembg-processor/requirements.txt
# Kaggle Enhanced REMBG Requirements
torch>=1.9.0
torchvision>=0.10.0
rembg>=2.0.0
opencv-python-headless>=4.5.0
pillow>=8.0.0
boto3>=1.20.0
numpy>=1.21.0
